backend/requirements.txt
# --- Core Flask Framework ---
Flask==2.3.3
gunicorn==21.2.0

# --- Database & ORM ---
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7

# --- Authentication & Security ---
Flask-JWT-Extended==4.5.2
bcrypt==4.0.1
cryptography==41.0.4 # A newer, more secure version
Flask-Cors==4.0.0

# --- API & Performance ---
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flasgger==0.9.7.1
redis==4.6.0

# --- Background Tasks ---
celery==5.3.1

# --- External Services & Utilities ---
requests==2.31.0
twilio==8.9.0
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0